requirements.txt
asgiref==3.8.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.3
django-import-export==3.3.7
django-mdeditor==0.1.20
django-unfold==0.21.1
et-xmlfile==1.1.0
Markdown==3.6
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.2.0
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
xlrd==2.0.1
xlwt==1.3.0
